pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "revops-copilot"
version = "0.1.0"
description = "RevOps Copilot — an agentic CLI for Private Equity portfolio deal management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pandas>=2.1",
    "pydantic>=2.5",
    "anthropic>=0.40",
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
]

[project.scripts]
revops = "src.cli:cli"

[tool.setuptools.packages.find]
include = ["src*"]
